refactorizes naming of blm values policy factory to be kebab case
