+++ api/apps/api/src/modules/projects/blm/project-blm.module.ts
@@ -10,7 +10,7 @@ import { ProjectBlmSaga } from './project-blm.saga';
 import { ChangeBlmRangeHandler } from './change-blm-range.handler';
 import { PlanningUnitAreaFetcher } from '@marxan-api/modules/projects/blm/planning-unit-area-fetcher';
 import { Project } from '@marxan-api/modules/projects/project.api.entity';
-import { BlmValuesPolicyFactory } from '@marxan-api/modules/projects/blm/BlmValuesPolicyFactory';
+import { BlmValuesPolicyFactory } from '@marxan-api/modules/projects/blm/blm-values-policy-factory';
 
 @Module({
   imports: [
